Refactor: Move generic middleware and multitoken middleware to own package
This creates a clearer separation in the code base.
