+++ ginjwt/jwt.go
@@ -2,7 +2,6 @@ package ginjwt
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"strings"
@@ -11,6 +10,8 @@ import (
 	"github.com/gin-gonic/gin"
 	"gopkg.in/square/go-jose.v2"
 	"gopkg.in/square/go-jose.v2/jwt"
+
+	"go.hollow.sh/toolbox/ginauth"
 )
 
 const (
@@ -63,40 +64,40 @@ func NewAuthMiddleware(cfg AuthConfig) (*Middleware, error) {
 
 // VerifyToken verifies a JWT token gotten from the gin.Context object against the given scopes.
 // This implements the GenericMiddleware interface
-func (m *Middleware) VerifyToken(c *gin.Context, scopes []string) (ClaimMetadata, error) {
+func (m *Middleware) VerifyToken(c *gin.Context, scopes []string) (ginauth.ClaimMetadata, error) {
 	authHeader := c.Request.Header.Get("Authorization")
 
 	if authHeader == "" {
-		return ClaimMetadata{}, NewAuthenticationError("missing authorization header, expected format: \"Bearer token\"")
+		return ginauth.ClaimMetadata{}, ginauth.NewAuthenticationError("missing authorization header, expected format: \"Bearer token\"")
 	}
 
 	authHeaderParts := strings.SplitN(authHeader, " ", expectedAuthHeaderParts)
 
 	if !(len(authHeaderParts) == expectedAuthHeaderParts && strings.ToLower(authHeaderParts[0]) == "bearer") {
-		return ClaimMetadata{}, NewAuthenticationError("invalid authorization header, expected format: \"Bearer token\"")
+		return ginauth.ClaimMetadata{}, ginauth.NewAuthenticationError("invalid authorization header, expected format: \"Bearer token\"")
 	}
 
 	rawToken := authHeaderParts[1]
 
 	tok, err := jwt.ParseSigned(rawToken)
 	if err != nil {
-		return ClaimMetadata{}, NewAuthenticationError("unable to parse auth token")
+		return ginauth.ClaimMetadata{}, ginauth.NewAuthenticationError("unable to parse auth token")
 	}
 
 	if tok.Headers[0].KeyID == "" {
-		return ClaimMetadata{}, NewAuthenticationError("unable to parse auth token header")
+		return ginauth.ClaimMetadata{}, ginauth.NewAuthenticationError("unable to parse auth token header")
 	}
 
 	key := m.getJWKS(tok.Headers[0].KeyID)
 	if key == nil {
-		return ClaimMetadata{}, NewInvalidSigningKeyError()
+		return ginauth.ClaimMetadata{}, NewInvalidSigningKeyError()
 	}
 
 	cl := jwt.Claims{}
 	sc := map[string]interface{}{}
 
 	if err := tok.Claims(key, &cl, &sc); err != nil {
-		return ClaimMetadata{}, NewAuthenticationError("unable to validate auth token")
+		return ginauth.ClaimMetadata{}, ginauth.NewAuthenticationError("unable to validate auth token")
 	}
 
 	err = cl.Validate(jwt.Expected{
@@ -105,7 +106,7 @@ func (m *Middleware) VerifyToken(c *gin.Context, scopes []string) (ClaimMetadata
 		Time:     time.Now(),
 	})
 	if err != nil {
-		return ClaimMetadata{}, NewTokenValidationError(err)
+		return ginauth.ClaimMetadata{}, ginauth.NewTokenValidationError(err)
 	}
 
 	var roles []string
@@ -119,7 +120,7 @@ func (m *Middleware) VerifyToken(c *gin.Context, scopes []string) (ClaimMetadata
 	}
 
 	if !hasScope(roles, scopes) {
-		return ClaimMetadata{}, NewAuthorizationError("not authorized, missing required scope")
+		return ginauth.ClaimMetadata{}, ginauth.NewAuthorizationError("not authorized, missing required scope")
 	}
 
 	var user string
@@ -130,7 +131,7 @@ func (m *Middleware) VerifyToken(c *gin.Context, scopes []string) (ClaimMetadata
 		user = cl.Subject
 	}
 
-	return ClaimMetadata{Subject: cl.Subject, User: user}, nil
+	return ginauth.ClaimMetadata{Subject: cl.Subject, User: user}, nil
 }
 
 // AuthRequired provides a middleware that ensures a request has authentication
@@ -142,7 +143,7 @@ func (m *Middleware) AuthRequired(scopes []string) gin.HandlerFunc {
 
 		cm, err := m.VerifyToken(c, scopes)
 		if err != nil {
-			abortBecauseOfError(c, err)
+			ginauth.AbortBecauseOfError(c, err)
 			return
 		}
 
@@ -160,7 +161,7 @@ func (m *Middleware) refreshJWKS() error {
 	defer resp.Body.Close()
 
 	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("%w: %s", ErrMiddlewareRemote, resp.Body)
+		return fmt.Errorf("%w: %s", ginauth.ErrMiddlewareRemote, resp.Body)
 	}
 
 	return json.NewDecoder(resp.Body).Decode(&m.cachedJWKS)
@@ -209,20 +210,4 @@ func GetSubject(c *gin.Context) string {
 // has already occurred. If authentication failed or there isn't a user an empty string is returned.
 func GetUser(c *gin.Context) string {
 	return c.GetString(contextKeyUser)
-}
-
-func abortBecauseOfError(c *gin.Context, err error) {
-	var authErr *AuthError
-
-	var validationErr *TokenValidationError
-
-	switch {
-	case errors.As(err, &validationErr):
-		c.AbortWithStatusJSON(validationErr.HTTPErrorCode, gin.H{"message": "invalid auth token", "error": validationErr.Error()})
-	case errors.As(err, &authErr):
-		c.AbortWithStatusJSON(authErr.HTTPErrorCode, gin.H{"message": authErr.Error()})
-	default:
-		// If we can't cast it, unauthorize anyway
-		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"message": err.Error()})
-	}
 }
